Fixed the modifier errors
Changed the order of execution
Nullcheck on spell before cast
Renamed the master in preparation for more singleton use
