+++ Assets/Scripts/Spells/Spell.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Creature;
+using Spells.Modifiers;
 using UnityEngine;
 
 namespace Spells
