+++ Assets/Scripts/Spells/SpellModifier.cs
@@ -3,24 +3,27 @@ using System;
 
 namespace Spells
 {
-    public enum ModifierTier
+    namespace Modifiers
     {
-        Tier1, // temp names
-        Tier2,
-        Tier3
-    }
-    
-    [Serializable]
-    public abstract class SpellModifier
-    {
-        public abstract string Name { get; }
-        public abstract string Description { get; }
-        public abstract ModifierTier Tier { get; }
+        public enum ModifierTier
+        {
+            Tier1, // temp names
+            Tier2,
+            Tier3
+        }
 
-        public abstract SpellStats ModifyStats(SpellStats currentStats);
+        [Serializable]
+        public abstract class SpellModifier
+        {
+            public abstract string Name { get; }
+            public abstract string Description { get; }
+            public abstract ModifierTier Tier { get; }
 
-        public abstract void Update(SpellProjectile projectile);
+            public abstract SpellStats ModifyStats(SpellStats currentStats);
 
-        public abstract void OnCollision(CreatureBase creature);
+            public abstract void Update(SpellProjectile projectile);
+
+            public abstract void OnCollision(CreatureBase creature);
+        }
     }
 }
